fixed styles
